requirements.txt
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
python-dateutil==2.8.2